exodus-android-app: Show appropriate chip colors
