+++ app/src/main/java/org/eu/exodus_privacy/exodusprivacy/fragments/appdetail/subfrags/ADTrackersFragment.kt
@@ -1,21 +1,19 @@
 package org.eu.exodus_privacy.exodusprivacy.fragments.appdetail.subfrags
 
-import android.content.res.ColorStateList
 import android.net.Uri
 import android.os.Bundle
 import android.view.View
 import androidx.browser.customtabs.CustomTabsIntent
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.google.android.material.chip.Chip
 import dagger.hilt.android.AndroidEntryPoint
 import org.eu.exodus_privacy.exodusprivacy.R
 import org.eu.exodus_privacy.exodusprivacy.databinding.FragmentADTrackersBinding
 import org.eu.exodus_privacy.exodusprivacy.fragments.appdetail.AppDetailViewModel
 import org.eu.exodus_privacy.exodusprivacy.fragments.trackers.model.TrackersRVAdapter
+import org.eu.exodus_privacy.exodusprivacy.utils.setExodusColor
 import javax.inject.Inject
 
 @AndroidEntryPoint
@@ -84,18 +82,5 @@ class ADTrackersFragment : Fragment(R.layout.fragment_a_d_trackers) {
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
-    }
-
-    private fun Chip.setExodusColor(size: Int) {
-        val colorRed = ContextCompat.getColor(context, R.color.colorRedLight)
-        val colorYellow = ContextCompat.getColor(context, R.color.colorYellow)
-        val colorGreen = ContextCompat.getColor(context, R.color.colorGreen)
-
-        val colorStateList = when (size) {
-            0 -> ColorStateList.valueOf(colorGreen)
-            in 1..4 -> ColorStateList.valueOf(colorYellow)
-            else -> ColorStateList.valueOf(colorRed)
-        }
-        this.chipBackgroundColor = colorStateList
     }
 }
